analysis

+++ xip.c
@@ -64,9 +64,9 @@ void rb_insert_node(struct rb_root *root, struct hash_map_addr *hash_map_addr_ne
 		parent = *entry_node;
 		hash_map_addr_entry = rb_entry(*entry_node, struct hash_map_addr, node);
 		
-		if(hash_map_addr_new->hashing < hash_map_addr_entry->hashing)
+		if((long)hash_map_addr_new->hashing < (long)hash_map_addr_entry->hashing)
 			entry_node = &(*entry_node)->rb_left;
-		else if(hash_map_addr_new->hashing > hash_map_addr_entry->hashing){
+		else if((long)hash_map_addr_new->hashing > (long)hash_map_addr_entry->hashing){
 			entry_node = &(*entry_node)->rb_right;
 		}	
 		else{
